+++ src/constants.ts
@@ -1,3 +1,3 @@
 export const defaultCaptureKey = '@@default-capture@@'
-export const specialFilterNames = ['html', 'node', 'text']
+export const specialFilterNames = ['html', 'outerHTML', 'node', 'text']
 export const universalSelector = '*'