+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## WIP
 
 * Temme now requires node >= 6. (Before is 7.6+)
-* Add built-in filter `get`.
+* Add built-in filter `get`
+* Add special filter `outerHTML`